utils/azure-api-tester/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
rich>=13.0.0
click>=8.1.0
